refactor: calendar components
